refactor(utils): share room quote snippet helper
Move buildQuoteSnippet into @sokosumi/utils so Core persist and web
composer preview use one implementation.
